SDK-138 Java SDK: Stabilization of FirehoseServiceTest
